git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@15203 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/USER-DPD/compute_dpd.cpp
@@ -36,9 +36,9 @@ ComputeDpd::ComputeDpd(LAMMPS *lmp, int narg, char **arg) :
   vector_flag = 1;
   size_vector = 5;
   extvector = 0;
-  
+
   vector = new double[size_vector];
-  
+
   if (atom->dpd_flag != 1) error->all(FLERR,"compute dpd requires atom_style with internal temperature and energies (e.g. dpd)");
 }
 
@@ -67,7 +67,7 @@ void ComputeDpd::compute_vector()
 
   dpdU = new double[size_vector];
 
-  for (int i = 0; i < size_vector; i++) dpdU[i] = double(0.0);
+  for (int i = 0; i < size_vector; i++) dpdU[i] = 0.0;
 
   for (int i = 0; i < nlocal; i++){
     if (mask[i] & groupbit){
@@ -78,7 +78,7 @@ void ComputeDpd::compute_vector()
       dpdU[4]++;
     }
   }
-    
+
   MPI_Allreduce(dpdU,vector,size_vector,MPI_DOUBLE,MPI_SUM,world);
 
   natoms = vector[4];
